+++ NEWS.md
@@ -1,3 +1,6 @@
+# traits4models 0.3.0
+* MeasurementMethod and AggregationLevel fields added
+
 # traits4models 0.2.4
 * Consistency check in function 'check_medfate_params()'
 * New trait definition 'PhotosyntheticPathway'